Create views with autolayout
Autolayout makes it much easier to create layouts that chance when
you resize, saving a lot of work. I'm not convinced that the
visual language for describing the relationships is the best
or the easiest. Expecting a period of harsh trial and error
followed by some level of mastery and a long, long plateau, as
with most foundation-level obj-c technology in OS X.