Don't allow users to open multiple new server windows
This was more difficult under ARC than I had originally envisioned.
For some reason, keeping a __strong reference to an NSWindow and
setting it to nil when we dismissed that window would cause a
crash consistently somewhere in - release.

Instead, I have a strong reference to an NSWindowController and a
flag to indicate when it's set or not. Kind of an ugly workaround,
but until I get a chance to learn more about this, I think
this will work.