Made default values for textfields easier to manage
I wanted to add a real default for the channels to join by default,
but I didn't want to write some more code that says

if (str.length == 0)
  [setValue: @"someDefaultValue"];

so GLGDefaultValueTextField will handle that for you. You just
ask for the stringValue and it gives you either the user input
or the default value.