(wip) Initial support for tabs.
I want to recreate the experience of the tab bar from either adium or chrome, and I
haven't found a control that supports ARC, so it looks like I'll be reinventing
the wheel for this.

The goal, in short is to have tabs that you can switch between easily, redraw the
chat area quickly, and make them look *good* and native. You should be able to
sort them, drag them out (and back in), close them, and generally do anything
that you could do with chrome or adium tabs.

When I'm done, it might be nice to pull this out and make it available as
a cocoapod, as an experience in giving back to the community.