Pull CoreData managed object context off App Delegate
Kept thinking about how complete insane this is. I suspect
it will be better if the interface is static, but this
should get us 90% of the way there.

The reason you wouldn't want to have this on the application
delegate is because anything that needs to save a record is
going to end up #importing the app delegate header, and likely
causing circular includes in the process.