Fix data model for server's channels
Man, what a nightmare. I set the ParentEntity on Channel to Server
at some point, and that was a mistake. That resulted in a lot of
weird looking records being saved because that property actually
is for inheritance, not for relations.

Now channels are optional on server, but hostname, port, etc are
all required. Name is required on channel and so is the server.