Fix "clipping" in tabview
If you hovered over any of the tabs, or used ctrl+tab to move
between them, you would notice there was some odd graphics
around the nsbezierpath that clips the view.

The reason for this was that we did not tell the superview to
redraw. It *is* nice that this redrawing can happen in basically
any order. I suppose as long as the subview isn't telling its
superview to redraw from within its drawRect routine, we won't
end up in an infinite loop or anything terrible like that.