Handle window resize
I really wanted to just use autolayout, but my autolayout-fu is not
strong enough to handle this simple layout. I just wanted to have
all of the three views (tabs, chat log, and text input be stacked
one of top of the other vertically and fill the entire view, but
that seemed to be difficult to implement, for some unknown reason.