Unbreak private messages
If you're sending a private message to another user, you don't prefix
it with a hash. All of this code needs to be broken out of the IRCBroker
class into something that can more sustainably parse a message and
result in some behavior being taken by the chatview and broker.