Don't display "nick has left" for yourself
I actually want to display this message, but only IFF the tab for that
channel is still open. This will require factoring out the messages such
that the chatview can know what the type of a message is and decide if it
wants to open a tab just to display a channel quit message.