Remove a hack for the current IRC server
Since you can connect to multiple servers (or even the same one
multiple times, it's important that we be able to differentiate
between which broker we are actually communicating with.

We don't close brokers if the last channel from that broker is
closed, but this needs to be tested more thoroughly first anyway.