Include quit message in channels when a user quits
Previously it would only display in the tab for the server
eg: chat.freenode.net

But now, if a user is in multiple channels and they quit, that message
will be displayed in each channel. I might find this annoying in practice,
so I might end up changing how the tabs are highlighted when this happens,
or including a setting to disable highlighting tabs when someone goes offline
or quits the channel.