Small message parsing refactoring
I was pairing with @habutai last night and she mentioned how it
seemed overly verbose to have all of this repeated code
where we are setting the properties on the IRCMessage we're
creating. Making a change to only setting it once at the end
is a little better, but not as big a win as I had hoped.

Seems like it would be ideal if a lot of this code could just be deleted,
but I'm not entirely sure what this could be replaced with. There is the
possibility of using NSScanner but I'm not entirely sure that would be
a better or more elegant solution