WIP: break out input and response parsing classes
Starting to think about how to test this helped inform me that some of
the design decisions I had made earlier were making this hard to test.

* the GLGIRCMessage class really only encapsulates some data, and doesn't have
any methods. In reality, the person creating those should only need to send
a message to a delegate, or its receiver
* The GLGIRCParser class served double duty for user input and response
parsing. These are separate concerns that can be broken out
* the GLGIRCParser was never instantiated, and only returned objects
that wrapped the state for the message parsed. Changing this so that
the relationship between how it parses messages is covered by a delegate
protocol makes it easier to test, and ensures that this contract is
honored at compile time. Also, these messages are simpler to read and
there is only one check for which message we received, instead of having
two or more.