Start refactoring IRCBroker and ChatView to use new input protocol
Thinking about how this work soon, I think I want to break apart some of
these objects so this object graph is less annoying to manage. Probably
breaking out a controller for the chatview would help a lot, but there's
also the issue of how the brokers for the chatview will get the text
needed to display back to the chatview.