Finish refactoring response and input parsers
Feeling better about this. Some failing tests that need some investigation
as far as what the correct format of these messages are, but this is already
a lot better. Now, if I could expand the tests to ensure that we use this
and then interpolate the actual <__nick__> and <__channel__> strings, we
should be golden.