Don't use an NSSet for drawing channel occupants sidebar
I believe this was my original solution when I discovered that you would
occasionally have the same name appear multiple times in the sidebar. A better
fix for this would be to use a data structure that is ordered and has distinct
objects (e.g.: duplicates are not allowed). NSSet is distinct, but unordered.

Perhaps the storage in the broker for the occupants **could** be an NSSet and it
could be converted to an Array as needed? That would make the broker code a lot
less fragile.