Remove one small hack and one larger one
One small hack : a boolean in AppDelegate so that we know when the
"connect to a new server" dialog is visible. The fix for this is to ask
the class of the contentview of the current visible window -- and then to
make it key and visible if it already is the current window.

One large hack : all kinds of nonsense in the NewServerView knew about its
superview. That shouldn't strictly be necessary. I suppose when I wrote this code
I was still learning a lot of AppKit, in particular how NSWindows and NSViews fit
together with ContentView and subViews. Now I realize that all of that code was
completely and entirely unnecessary.